Trait backed_data::array::container::Container

pub trait Container: RefIter<Self::Data> + MutIter<Self::Data> {
    type Data;
    type Ref<'b>: AsRef<Self::Data> + Deref<Target = Self::Data> + StableDerefFeatureOpt
       where Self: 'b;
    type Mut<'b>: AsMut<Self::Data> + DerefMut<Target = Self::Data> + StableDerefFeatureOpt
       where Self: 'b;
    type RefSlice<'b>: AsRef<[Self::Data]> + Deref<Target = [Self::Data]> + StableDerefFeatureOpt
       where Self: 'b;
    type MutSlice<'b>: AsMut<[Self::Data]> + DerefMut<Target = [Self::Data]> + StableDerefFeatureOpt
       where Self: 'b;

    // Required methods
    fn c_get(&self, index: usize) -> Option<Self::Ref<'_>>;
    fn c_get_mut(&mut self, index: usize) -> Option<Self::Mut<'_>>;
    fn c_len(&self) -> usize;
    fn c_ref(&self) -> Self::RefSlice<'_>;
    fn c_mut(&mut self) -> Self::MutSlice<'_>;
}

Available on crate feature array only.

Generic wrapper for any container type.

This is implemented over standard Rust containers (Vec, boxed slices, etc.) in addition to more exotic types. &[T] is insufficiently generic for types that return a ref handle to T, instead of &T directly, so this allows for more complex container types.

Methods are prepended with c_* to avoid namespace conflicts.

Required Associated Types

type Data

The data container entries give references to.

type Ref<'b>: AsRef<Self::Data> + Deref<Target = Self::Data> + StableDerefFeatureOpt where Self: 'b

type Mut<'b>: AsMut<Self::Data> + DerefMut<Target = Self::Data> + StableDerefFeatureOpt where Self: 'b

type RefSlice<'b>: AsRef<[Self::Data]> + Deref<Target = [Self::Data]> + StableDerefFeatureOpt where Self: 'b

type MutSlice<'b>: AsMut<[Self::Data]> + DerefMut<Target = [Self::Data]> + StableDerefFeatureOpt where Self: 'b

Required Methods

fn c_get(&self, index: usize) -> Option<Self::Ref<'_>>

fn c_get_mut(&mut self, index: usize) -> Option<Self::Mut<'_>>

fn c_len(&self) -> usize

fn c_ref(&self) -> Self::RefSlice<'_>

fn c_mut(&mut self) -> Self::MutSlice<'_>

Object Safety

This trait is not object safe.