backed_data::array::container
Trait StableDerefFeatureOpt

pub trait StableDerefFeatureOpt: StableDeref { }

Available on crate features array and unsafe_array only.

Requires stable_deref_trait::StableDeref when unsafe code is enabled.

If unsafe code is not enabled, this implements for everything. This hackery gets around a trap between borrow checker limitations and forbid(unsafe_code) limitations.