Struct backed_data::array::BackedArray

pub struct BackedArray<K, E> { /* private fields */ }

Available on crate feature array only.

Array stored as multiple arrays on disk.

Associates each access with the appropriate disk storage, loading it into memory and returning the value. Subsequent accesses will use the in-memory store. Use Self::clear_memory or Self::shrink_to_query to move the cached sub-arrays back out of memory.

All mutation must be through the iterator types, which guarantee flushes to underlying storage.

Some container types provide handles instead of direct &T references. They can only be interfaced through the *generic_* methods, which add overhead to return the references with their handles. Types that dereference directly to &[T] (e.g. Vec) should not use the generic methods.

Generics

K: Container that holds usize values.

E: Container that holds BackedEntry values.

Implementations

impl<K: Container<Data = usize>, E: BackedEntryContainerNestedAsyncRead> BackedArray<K, E>

Read implementations

pub async fn a_generic_get( &self, idx: usize, ) -> Result<impl Deref<Target = E::InnerData> + '_, BackedArrayError<E::AsyncReadError>>

Available on crate features async and unsafe_array only.

Async version of Self::generic_get.

impl<K: Container<Data = usize>, E: BackedEntryContainerNestedAsyncRead> BackedArray<K, E>

pub fn generic_stream(&self) -> BackedArrayFutIterGeneric<'_, K, E>

Available on crate features async and unsafe_array only.

Future iterator over each backed item.

Can be converted to a Stream with futures::stream::iter. This is not a stream by default because Stream does not have the iterator methods that allow efficient implementations.

Use Self::stream_send for Send bounds.

pub fn generic_stream_send( this: &Arc<Self>, ) -> BackedArrayFutIterGenericSend<K, E>
where K: Send + Sync + 'static, for<'b> K::Ref<'b>: Send + Sync, E: Send + Sync + 'static, E::Coder: Send + Sync, E::Disk: Send + Sync, <E::Disk as AsyncReadDisk>::ReadDisk: Send + Sync, E::OnceWrapper: Send + Sync, for<'b> E::Ref<'b>: Send + Sync, for<'b> <E::Unwrapped as Container>::Ref<'b>: Send + Sync,

Available on crate features async and unsafe_array only.

Version of Self::stream with Send bounds.

If this type owns the disk (all library disks are owned), wrapping in an Arc and calling with BackedArray::stream_send(&this) satisfies all of the bounds.

pub fn generic_chunk_stream( &self, ) -> impl Iterator<Item = impl Future<Output = Result<impl Deref<Target = &E::Unwrapped>, E::AsyncReadError>>>

Available on crate features async and unsafe_array only.

Future iterator over each chunk.

Can be converted to a Stream with futures::stream::iter. This is not a stream by default because Stream does not have the iterator methods that allow efficient implementations.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsRef<[E::Data]> + BackedEntryContainerNestedAsyncRead, E::Unwrapped: AsRef<[E::InnerData]>,

Read implementations

pub async fn a_get( &self, idx: usize, ) -> Result<&E::InnerData, BackedArrayError<E::AsyncReadError>>

Available on crate feature async only.

Async version of Self::get.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsRef<[E::Data]> + BackedEntryContainerNestedAsyncRead, E::Unwrapped: AsRef<[E::InnerData]>,

pub fn stream(&self) -> BackedArrayFutIter<'_, K, E>

Available on crate feature async only.

Future iterator over each backed item.

Can be converted to a Stream with futures::stream::iter. This is not a stream by default because Stream does not have the iterator methods that allow efficient implementations.

Use Self::stream_send for + Send bounds.

pub fn stream_send(this: &Arc<Self>) -> BackedArrayFutIterSend<K, E>
where K: Send + Sync + 'static, for<'b> K::Ref<'b>: Send + Sync, E: Send + Sync + 'static, E::Coder: Send + Sync, E::Disk: Send + Sync, <E::Disk as AsyncReadDisk>::ReadDisk: Send + Sync, E::OnceWrapper: Send + Sync, for<'b> E::Ref<'b>: Send + Sync, for<'b> <E::Unwrapped as Container>::Ref<'b>: Send + Sync,

Available on crate feature async only.

Version of Self::stream with + Send bounds.

If this type owns the disk (all library disks are owned), wrapping in an Arc and calling with BackedArray::stream_send(&this) satisfies all of the bounds.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsRef<[E::Data]> + BackedEntryContainerNestedAsyncRead,

pub fn chunk_stream( &self, ) -> impl Iterator<Item = impl Future<Output = Result<&E::Unwrapped, E::AsyncReadError>>>

Available on crate feature async only.

Future iterator over each chunk.

Can be converted to a Stream with futures::stream::iter. This is not a stream by default because Stream does not have the iterator methods that allow efficient implementations.

impl<K, E> BackedArray<K, E>
where K: From<Vec<K::Data>> + Container<Data = usize>, E: From<Vec<BackedEntryAsync<E::Unwrapped, E::Disk, E::Coder>>> + BackedEntryContainerNestedAsyncWrite, E::Disk: Clone, E::Coder: Clone,

pub async fn a_from_containers<C, I>( iter: I, disk: &E::Disk, coder: &E::Coder, ) -> Result<Self, <E::Coder as AsyncEncoder<<E::Disk as AsyncWriteDisk>::WriteDisk>>::Error>
where C: Into<E::Unwrapped>, I: IntoIterator<Item = C>,

Available on crate feature async only.

Async version of Self::from_containers.

Executes between 1 and [# iterator entries] write futures concurrently. Keeps all data in memory.

impl<K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedAsyncWrite + ResizingContainer> BackedArray<K, E>

pub async fn a_append<U: Into<E::Unwrapped>>( &mut self, values: U, backing_store: E::Disk, coder: E::Coder, ) -> Result<&mut Self, E::AsyncWriteError>

Available on crate feature async only.

Adds new values by writing them to the backing store.

Does not keep the values in memory.

Example

#[cfg(feature = "async_bincode")] {
use backed_data::{
    array::async_impl::AsyncVecBackedArray,
    entry::{
        disks::Plainfile,
        formats::AsyncBincodeCoder,
    }
};
use std::fs::{File, create_dir_all, remove_dir_all, OpenOptions};

use tokio::runtime::Builder;

let rt = Builder::new_current_thread().build().unwrap();

let FILENAME_BASE = std::env::temp_dir().join("example_async_array_append");
let values = ([0, 1, 1],
    [2, 3, 5]);

create_dir_all(FILENAME_BASE.clone()).unwrap();
let file_0 = FILENAME_BASE.clone().join("_0");
let file_1 = FILENAME_BASE.join("_1");
let mut array: AsyncVecBackedArray<u32, Plainfile, _> = AsyncVecBackedArray::default();
rt.block_on(array.a_append(values.0, file_0.into(), AsyncBincodeCoder::default()));
rt.block_on(array.a_append(values.1, file_1.into(), AsyncBincodeCoder::default()));

assert_eq!(*rt.block_on(array.a_get(4)).unwrap(), 3);
remove_dir_all(FILENAME_BASE).unwrap();
}

pub async fn a_append_memory<U: Into<E::Unwrapped>>( &mut self, values: U, backing_store: E::Disk, coder: E::Coder, ) -> Result<&mut Self, E::AsyncWriteError>

Available on crate feature async only.

Self::append, but keeps values in memory.

Example

#[cfg(feature = "async_bincode")] {
    use backed_data::{
        array::async_impl::AsyncVecBackedArray,
        entry::{
            disks::Plainfile,
            formats::AsyncBincodeCoder,
        },
    };
    use std::fs::{File, remove_file, OpenOptions};

    use tokio::runtime::Builder;

    let rt = Builder::new_current_thread().build().unwrap();

    let FILENAME = std::env::temp_dir().join("example_async_array_append_memory");
    let values = ([0, 1, 1],
        [2, 3, 5]);

    let mut array: AsyncVecBackedArray<u32, Plainfile, _> = AsyncVecBackedArray::default();
    rt.block_on(array.a_append_memory(values.0, FILENAME.clone().into(), AsyncBincodeCoder::default()));

    // Overwrite file, making disk pointer for first array invalid
    rt.block_on(array.a_append_memory(values.1, FILENAME.clone().into(), AsyncBincodeCoder::default()));

    assert_eq!(*rt.block_on(array.a_get(0)).unwrap(), 0);
    remove_file(FILENAME).unwrap();
}

impl<K: Container<Data = usize>, E: BackedEntryContainerNestedRead> BackedArray<K, E>

Read implementations

pub fn generic_get( &self, idx: usize, ) -> Result<impl Deref<Target = E::InnerData> + '_, BackedArrayError<E::ReadError>>

Available on crate feature unsafe_array only.

Self::get that works on containers that don’t directly dereference to slices.

impl<K: Container<Data = usize>, E: BackedEntryContainerNestedRead> BackedArray<K, E>

pub fn generic_iter(&self) -> BackedArrayGenericIter<'_, K, E>

Available on crate feature unsafe_array only.

Self::iter that works on containers that don’t directly dereference to slices.

pub fn generic_chunk_iter( &self, ) -> impl Iterator<Item = Result<impl Deref<Target = &E::Unwrapped>, E::ReadError>>

Available on crate feature unsafe_array only.

Self::chunk_iter that works on containers that don’t directly dereference to slices.

impl<K: Container<Data = usize>, E: BackedEntryContainerNestedAll> BackedArray<K, E>

pub fn generic_iter_mut<'a>(&'a mut self) -> BackedArrayGenericIterMut<'a, K, E>
where E::Unwrapped: 'a, E::ReadError: 'a,

Available on crate feature unsafe_array only.

Self::iter_mut that works on containers that don’t directly dereference to slices.

pub fn generic_chunk_mut_iter( &mut self, ) -> impl Iterator<Item = Result<impl Deref<Target = BackedEntryMut<'_, BackedEntry<E::OnceWrapper, E::Disk, E::Coder>>>, E::ReadError>>

Available on crate feature unsafe_array only.

Self::chunk_mut_iter that works on containers that don’t directly dereference to slices.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsRef<[E::Data]> + BackedEntryContainerNestedRead, E::Unwrapped: AsRef<[E::InnerData]>,

Read implementations

pub fn get( &self, idx: usize, ) -> Result<&E::InnerData, BackedArrayError<E::ReadError>>

Return a value (potentially loading its backing array).

Backing arrays stay in memory until freed by a mutable function.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsRef<[E::Data]> + BackedEntryContainerNestedRead, E::Unwrapped: AsRef<[E::InnerData]>,

pub fn iter(&self) -> BackedArrayIter<'_, K, E>

Iterates over each backed item.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsRef<[E::Data]> + BackedEntryContainerNestedRead,

pub fn chunk_iter( &self, ) -> impl Iterator<Item = Result<&E::Unwrapped, E::ReadError>>

Returns underlying chunks in order.

This will load each chunk before providing it. All chunks loaded earlier in the iteration will remain loaded.

impl<K, E> BackedArray<K, E>
where K: AsRef<[usize]> + Container<Data = usize>, E: AsMut<[E::Data]> + BackedEntryContainerNestedAll, E::Unwrapped: AsMut<[E::InnerData]>,

pub fn iter_mut<'a>(&'a mut self) -> BackedArrayIterMut<'a, E>
where E::Unwrapped: 'a, E::ReadError: 'a,

Available on crate feature unsafe_array only.

Returns BackedArrayIterMut, which automatically tracks and forces flushes when mutated values are dropped. Both the iterator and its handles also provide manually callable flush methods.

impl<K: Container<Data = usize>, E> BackedArray<K, E>
where E: AsMut<[E::Data]> + BackedEntryContainerNestedAll,

pub fn chunk_mut_iter( &mut self, ) -> impl Iterator<Item = Result<BackedEntryMut<'_, BackedEntry<E::OnceWrapper, E::Disk, E::Coder>>, E::ReadError>>

Provides mutable handles to underlying chunks, using BackedEntryMut.

See Self::chunk_iter for the immutable iterator.

impl<K, E: BackedEntryContainerNested> BackedArray<K, E>

pub fn clear_memory(&mut self)

Move all backing arrays out of memory.

pub fn clear_chunk(&mut self, idx: usize)

Move the chunk at idx out of memory.

impl<K: Container<Data = usize>, E: BackedEntryContainerNested> BackedArray<K, E>

pub fn loaded_len(&self) -> usize

Returns the number of items currently loaded into memory.

To get memory usage on constant-sized items, multiply this value by the per-item size.

impl<K: Container<Data = usize>, E: BackedEntryContainerNested> BackedArray<K, E>

pub fn shrink_to_query(&mut self, idxs: &[usize])

Removes all backing stores not needed to hold idxs in memory.

impl<K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedWrite + ResizingContainer> BackedArray<K, E>

Write implementations

pub fn append<U: Into<E::Unwrapped>>( &mut self, values: U, backing_store: E::Disk, coder: E::Coder, ) -> Result<&mut Self, E::WriteError>

Adds new values by writing them to the backing store.

Does not keep the values in memory.

Example

#[cfg(feature = "bincode")] {
use backed_data::{
    array::VecBackedArray,
    entry::{
        disks::Plainfile,
        formats::BincodeCoder,
    }
};
use std::fs::{File, create_dir_all, remove_dir_all, OpenOptions};

let FILENAME_BASE = std::env::temp_dir().join("example_array_append");
let values = ([0, 1, 1],
    [2, 3, 5]);

create_dir_all(FILENAME_BASE.clone()).unwrap();
let file_0 = FILENAME_BASE.clone().join("_0");
let file_1 = FILENAME_BASE.join("_1");
let mut array: VecBackedArray<u32, Plainfile, _> = VecBackedArray::default();
array.append(values.0, file_0.into(), BincodeCoder::default());
array.append(values.1, file_1.into(), BincodeCoder::default());

assert_eq!(*array.get(4).unwrap(), 3);
remove_dir_all(FILENAME_BASE).unwrap();
}

pub fn append_memory<U: Into<E::Unwrapped>>( &mut self, values: U, backing_store: E::Disk, coder: E::Coder, ) -> Result<&mut Self, E::WriteError>

Self::append, but keeps values in memory.

Example

#[cfg(feature = "bincode")] {
    use backed_data::{
        array::VecBackedArray,
        entry::{
            disks::Plainfile,
            formats::BincodeCoder,
        },
    };
    use std::fs::{File, remove_file, OpenOptions};

    let FILENAME = std::env::temp_dir().join("example_array_append_memory");
    let values = ([0, 1, 1],
        [2, 3, 5]);

    let mut array: VecBackedArray<u32, Plainfile, _> = VecBackedArray::default();
    array.append_memory(values.0, FILENAME.clone().into(), BincodeCoder::default());

    // Overwrite file, making disk pointer for first array invalid
    array.append_memory(values.1, FILENAME.clone().into(), BincodeCoder::default());

    assert_eq!(*array.get(0).unwrap(), 0);
    remove_file(FILENAME).unwrap();
}

impl<K: ResizingContainer<Data = usize>, E: BackedEntryContainerNested + ResizingContainer> BackedArray<K, E>

pub fn remove(&mut self, entry_idx: usize) -> &mut Self

Removes an entry with the internal index, shifting ranges.

The getter functions can be used to identify the target index.

impl<K: ResizingContainer<Data = usize>, E: ResizingContainer> BackedArray<K, E>

pub fn merge(&mut self, rhs: Self) -> &mut Self

Move entries in rhs to self.

impl<K, E> BackedArray<K, E>
where K: From<Vec<K::Data>> + Container<Data = usize>, E: From<Vec<BackedEntry<E::OnceWrapper, E::Disk, E::Coder>>> + BackedEntryContainerNestedWrite, E::Disk: Clone, E::Coder: Clone,

pub fn from_containers<C, I>( iter: I, disk: &E::Disk, coder: &E::Coder, ) -> Result<Self, <E::Coder as Encoder<<E::Disk as WriteDisk>::WriteDisk>>::Error>
where C: Into<E::Unwrapped>, I: IntoIterator<Item = C>,

Builds self from an iterator of containers, keeping them all loaded in memory.

impl<K, E> BackedArray<K, E>
where K: From<Vec<K::Data>> + Container<Data = usize>, E: From<Vec<E::Data>> + BackedEntryContainerNestedWrite, E::Disk: Clone, E::Coder: Clone,

pub fn from_backing<I>(iter: I) -> Option<Self>
where I: IntoIterator<Item = (usize, E::Data)>,

Builds self from existing backings.

The (len, backing) pairs must be equal length. Otherwise this will return None.

Parameters

iter: Iterator of (len, backing) pairs.

impl<K, E> BackedArray<K, E>

Auto-generated by derive_getters::Getters.

pub fn key_starts(&self) -> &K

Get field key_starts from instance of BackedArray.

pub fn key_ends(&self) -> &K

Get field key_ends from instance of BackedArray.

pub fn entries(&self) -> &E

Get field entries from instance of BackedArray.

impl<K: Default, E: Default> BackedArray<K, E>

pub fn new() -> Self

impl<K: Container<Data = usize>, E> BackedArray<K, E>

pub fn len(&self) -> usize

Total size of stored data.

impl<K, E: Container> BackedArray<K, E>

pub fn chunks_len(&self) -> usize

Number of underlying chunks.

pub fn raw_chunks(&mut self) -> impl Iterator<Item: AsMut<E::Data>> + '_

Access to the underlying chunks, without loading data.

pub fn is_empty(&self) -> bool