Struct backed_data::directory::DirectoryBackedArray

pub struct DirectoryBackedArray<K, E> {
    pub array: BackedArray<K, E>,
    /* private fields */
}

Available on crate feature directory only.

BackedArray wrapper for automatically creating new entries in a directory.

The array can be accessed directly when the wrapper methods are insufficient. The wrapper methods are sufficient for most operations.

Fields
array: BackedArray<K, E>
Implementations

impl<K, E> DirectoryBackedArray<K, E>
where K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedAll + ResizingContainer, E::Disk: AsRef<Path>,

pub async fn a_remove(&mut self, entry_idx: usize) -> Result<&mut Self, Error>

Available on runtime only.

pub async fn a_delete(self) -> Result<(), Error>

Available on runtime only.

impl<K, E> DirectoryBackedArray<K, E>
where K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedAsyncWrite + ResizingContainer, E::Coder: Default, E::Disk: From<PathBuf>,

pub async fn a_append<U: Into<E::Unwrapped>>( &mut self, values: U, ) -> Result<&mut Self, E::AsyncWriteError>

Available on runtime only.

pub async fn a_append_memory<U: Into<E::Unwrapped>>( &mut self, values: U, ) -> Result<&mut Self, E::AsyncWriteError>

Available on runtime only.

impl<K, E> DirectoryBackedArray<K, E>
where K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedAsyncWrite + ResizingContainer, E::Disk: AsRef<Path>,

pub async fn a_append_dir(&mut self, rhs: Self) -> Result<&mut Self, Error>

Available on runtime only.

impl<K, E> DirectoryBackedArray<K, E>
where K: Default, E: Default,

pub async fn a_new(directory_root: PathBuf) -> Result<Self>

Available on runtime only.

Creates a new directory at directory_root.

directory_root: Valid read/write directory on system

impl<K, E> DirectoryBackedArray<K, E>
where E: BackedEntryContainerNested, E::Disk: AsRef<Path> + From<PathBuf>,

pub async fn a_move_root(&mut self, new_root: PathBuf) -> Result<&mut Self>

Available on runtime only.

Moves the directory to a new location wholesale.

impl<K, E> DirectoryBackedArray<K, E>
where E: BackedEntryContainerNestedAsync, E::Disk: From<PathBuf>,

pub fn a_next_target(&self) -> E::Disk

Available on runtime only.

impl<K, E> DirectoryBackedArray<K, E>
where K: Send + Sync + Serialize, E: Send + Sync + Serialize + BackedEntryContainerNestedAsyncWrite, E::Disk: From<PathBuf>,

pub async fn a_save<C>(&self, coder: &C) -> Result<&Self, C::Error>
where C: AsyncEncoder<<E::Disk as AsyncWriteDisk>::WriteDisk, Error: From<Error>, T = Self>,

Available on runtime only.

Async version of Self::save.

impl<K, E> DirectoryBackedArray<K, E>
where K: Send + Sync + for<'de> Deserialize<'de>, E: Send + Sync + for<'de> Deserialize<'de> + BackedEntryContainerNestedAsyncRead, E::Disk: From<PathBuf>,

pub async fn a_load<P: AsRef<Path>, C>( root: P, coder: &C, ) -> Result<Self, C::Error>
where C: AsyncDecoder<<E::Disk as AsyncReadDisk>::ReadDisk, Error: From<Error>, T = Self>,

Available on runtime only.

Async version of Self::load.

impl<K, E> DirectoryBackedArray<K, E>
where K: Default, E: Default,

pub fn deconstruct( self, ) -> (Arc<BackedArray<K, E>>, impl Fn(Arc<BackedArray<K, E>>) -> Option<Self>)

Available on runtime only.

Splits this into parts, for functions that need Arc<BackedArray>.

Return Tuple

0: This backing array, wrapped in an Arc.

1: A function that rebuilds self from the Arc (fails if reference is not unique).

impl<K, E> DirectoryBackedArray<K, E>

pub fn get_directory_root(&self) -> &PathBuf

impl<K, E> DirectoryBackedArray<K, E>
where K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedAll + ResizingContainer, E::Disk: AsRef<Path>,

pub fn remove(&mut self, entry_idx: usize) -> Result<&mut Self, Error>

pub fn delete(self) -> Result<(), Error>

impl<K, E> DirectoryBackedArray<K, E>
where K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedWrite + ResizingContainer, E::Coder: Default, E::Disk: From<PathBuf>,

pub fn append<U: Into<E::Unwrapped>>( &mut self, values: U, ) -> Result<&mut Self, E::WriteError>

pub fn append_memory<U: Into<E::Unwrapped>>( &mut self, values: U, ) -> Result<&mut Self, E::WriteError>

impl<K, E> DirectoryBackedArray<K, E>
where K: ResizingContainer<Data = usize>, E: BackedEntryContainerNestedWrite + ResizingContainer, E::Disk: AsRef<Path>,

pub fn append_dir(&mut self, rhs: Self) -> Result<&mut Self, Error>

Append another self of the same type.

This will move all entries of rhs into self’s root, and clean up the rhs directory if empty after move.

pub fn append_array( &mut self, rhs: BackedArray<K, E>, ) -> Result<&mut Self, Error>

Append an existing BackedArray.

This will move all entries of rhs into self’s root.

impl<K, E> DirectoryBackedArray<K, E>
where K: Default, E: Default,

pub fn new(directory_root: PathBuf) -> Result<Self>

Creates a new directory at directory_root.

directory_root: Valid read/write directory on system

impl<K, E> DirectoryBackedArray<K, E>
where E: BackedEntryContainerNested, E::Disk: AsRef<Path> + From<PathBuf>,

pub fn update_root<P: AsRef<Path>>(&mut self, new_root: P) -> &mut Self

Updates the root of the directory backed array.

Does not move any files or directories, just changes the stored root.

pub fn move_root(&mut self, new_root: PathBuf) -> Result<&mut Self>

Moves the directory to a new location wholesale.

impl<K, E> DirectoryBackedArray<K, E>

pub fn from_existing_array( array: BackedArray<K, E>, directory_root: PathBuf, ) -> Self

impl<K, E> DirectoryBackedArray<K, E>
where K: Serialize, E: Serialize + BackedEntryContainerNestedWrite, E::Disk: From<PathBuf>,

pub fn save<C>(&self, coder: &C) -> Result<&Self, C::Error>
where C: Encoder<<E::Disk as WriteDisk>::WriteDisk, Error: From<Error>, T = Self>,

Save self at DIRECTORY_ROOT/meta.dat.

As self implements serialize and deserialize, it does not need to be saved with this method. However, this provides a standard location and format for Self::load to utilize.

impl<K, E> DirectoryBackedArray<K, E>
where K: for<'de> Deserialize<'de>, E: for<'de> Deserialize<'de> + BackedEntryContainerNestedRead, E::Disk: From<PathBuf>,

pub fn load<P: AsRef<Path>, C>(root: P, coder: &C) -> Result<Self, C::Error>
where C: Decoder<<E::Disk as ReadDisk>::ReadDisk, Error: From<Error>, T = Self>,

Load self from DIRECTORY_ROOT/meta.dat.

As self implements serialize and deserialize, it does not need to be loaded with this method. However, this uses a standard location and format from a Self::save call.

impl<K, E> DirectoryBackedArray<K, E>
where E: BackedEntryContainerNested, E::Disk: From<PathBuf>,

pub fn next_target(&self) -> E::Disk