Module backed_data::entry::adapters

Available on crate feature async only.

Defines adapters that allow synchronous disks and/or format encoders to be used asynchronously.

Structs

DecodeBg

BlockingFn that runs a synchronous decoder.

EncodeBg

BlockingFn that runs a synchronous encoder into a write sink.

SyncAsAsync

Adapts a synchronous disk to be asynchronous.

SyncAsAsyncRead

Runs a sync Read on a background thread for nonblocking AsyncRead.

SyncAsAsyncReadBg

BlockingFn that spawns a background reader.

SyncAsAsyncWrite

SyncAsAsyncWriteBg

BlockingFn that spawns a background writer.

SyncCoderAsAsync

Adapts a synchronous coder (with synchronous disks) to be asynchronous.

SyncCoderAsyncDisk

Adapts a synchronous coder and asynchronous disk to be asynchronous.