Struct backed_data::entry::adapters::SyncCoderAsyncDisk

pub struct SyncCoderAsyncDisk<C, D, RF, EF, DR, ER> { /* private fields */ }

Available on crate feature async only.

Adapts a synchronous coder and asynchronous disk to be asynchronous.

The function handle is used to spawn coderโs (de)serialization on another thread.

When this struct is deserialized, the handle is uninitialized. The handle needs to be initialized to use the AsyncEncoder and AsyncDecoder implementations.

Implementations

impl<C, D, RF, EF, DR, ER> SyncCoderAsyncDisk<C, D, RF, EF, DR, ER>

pub fn new(coder: C, read_handle: RF, encode_handle: EF) -> Self

pub fn set_handles(&mut self, read_handle: RF, encode_handle: EF) -> &mut Self