backed_data::entry

Module disks

Defines the underlying storage for BackedEntry.

The types Plainfile, WriteUnbuffered, Unbuffered, and Mmap are completely interchangeable. A disk written to by one can be read by another.

Re-exports

pub use zstd::ZstdDisk;zstd or async_zstd

pub use encrypted::Encrypted;encrypted

pub use encrypted::SecretVecWrapper;encrypted

pub use mmap::Mmap;mmap_impl

Modules

custom

Convenience Disk define structs.

encryptedencrypted

Defines memory protected storage that encrypts to disk.

mmapmmap_impl

Defines the Mmap type to wrap mmap.

zstdzstd or async_zstd

ZSTD encryption and decryption.

Structs

Networknetwork

Represents some inbound network request.

Plainfile

A regular file entry.

Unbuffered

Plainfile, but with no buffering at all.

WriteUnbuffered

Plainfile, but with no write buffering.

Traits

AsyncReadDiskasync

Produces storage that can be read from asynchronously.

AsyncWriteDiskasync

Produces storage that can be written to asynchronously.

ReadDisk

Produces storage that can be read from synchronously.

WriteDisk

Produces storage that can be written to synchronously.

Functions

default_clientnetwork

Clones out a new client from an internal static pool.

Type Aliases

AsyncErrorruntime

AsyncFileruntime