Struct backed_data::entry::disks::encrypted::Encrypted

pub struct Encrypted<'a, B> { /* private fields */ }

Available on crate feature encrypted only.

A resource encrypted with Aes256Gcm on some underlying disk.

Secrets and intermediate processing are stored in SecretBox, which minimizes the risk of memory being snooped (see secrets for details). This does not protect the code read from this disk, so read into SecretVecWrapper to keep memory guards in place beyond encryption and decryption.

The secrets are excluded from serialization and deserialization, being set with garbage data on deserialization. Use Self::set_key and Self::set_nonce to initialize secrets after deserialization.

Since encryption uses an internal buffer for reading/writing, Unbuffered is more efficient than Plainfile.

Implementations

impl<B> Encrypted<'_, B>

pub fn new<K, N>(inner: B, key: Option<K>, nonce: Option<N>) -> Self
where K: AsRef<[u8; 32]>, N: AsRef<[u8; 12]>,

Create a new Encrypted.

Parameters

inner: Underlying disk written to / read from

key: Aes256Gcm key. Generated with OsRng if not provided.

nonce: Aes256Gcm nonce. Generated with OsRng if not provided.

pub fn new_random(inner: B) -> Self

To avoid explicit typing every time with Self::new and Nones.

pub fn set_key(&mut self, key: &SecretBox<[u8; 32]>) -> &mut Self

pub fn set_nonce(&mut self, nonce: &SecretBox<[u8; 12]>) -> &mut Self

pub fn get_key(&self) -> impl Deref<Target = [u8; 32]> + '_

pub fn get_nonce(&self) -> impl Deref<Target = [u8; 12]> + '_