backed_data::entry::disks::encrypted

Struct SecretVecWrapper

pub struct SecretVecWrapper<T: Bytes + ?Sized>(pub SecretVec<T>);

Available on crate feature encrypted only.

Wraps secrets::SecretVec.

Implements Serialize and Deserialize, but serialization and deserialization leak information, so be careful. Using Encrypted as the backing store can make sure protection is kept during serialization, and the disk only sees encrypted data.

Tuple Fields
0: SecretVec<T>