Trait backed_data::entry::formats::Encoder

pub trait Encoder<Target: ?Sized> {
    type Error: From<Error>;
    type T: Serialize + ?Sized;

    // Required method
    fn encode(&self, data: &Self::T, target: Target) -> Result<(), Self::Error>;
}

A format encoder that can be used synchronously.

Implementation

It is the responsibility of trait implementors to call [WriteExt::flush] to ensure encodings write out to disk. Any implementor that does not to call these may fail to put all data on disk, making the encoding invalid and causing read failure.

Required Associated Types

type Error: From<Error>

type T: Serialize + ?Sized

Required Methods

fn encode(&self, data: &Self::T, target: Target) -> Result<(), Self::Error>

Fully write out formatted data to a target disk.