Struct backed_data::entry::BackedEntry

pub struct BackedEntry<T, Disk, Coder> { /* private fields */ }

Entry kept on some backing storage, loaded into memory on request.

Use a heap pointer type or BackedEntryBox, otherwise this will occupy the full type size even when unloaded.

Generics

T: a Once wrapper over the type to store.

Disk: an implementor of ReadDisk and/or WriteDisk.

Coder: an implementor of Encoder and/or Decoder.

See BackedEntryBox for a heap-storing version.

See BackedEntryArr for an array-specialized version.

See BackedEntryAsync with feature async enabled for async.

Implementations

impl<T: Serialize + Send + Sync, Disk: AsyncWriteDisk, Coder> BackedEntry<OnceCell<T>, Disk, Coder>
where Coder: AsyncEncoder<Disk::WriteDisk, T = T>,

pub async fn a_update(&mut self) -> Result<(), Coder::Error>

Available on crate feature async only.

See Self::update.

pub async fn a_write(&mut self, new_value: T) -> Result<(), Coder::Error>

Available on crate feature async only.

See Self::write.

impl<T: for<'de> Deserialize<'de> + Send + Sync, Disk: AsyncReadDisk, Coder> BackedEntry<OnceCell<T>, Disk, Coder>
where Coder: AsyncDecoder<Disk::ReadDisk, T = T>,

pub async fn a_load(&self) -> Result<&T, Coder::Error>

Available on crate feature async only.

See Self::load.

impl<T: Serialize + Send + Sync, Disk: AsyncWriteDisk, Coder> BackedEntry<OnceCell<T>, Disk, Coder>
where Coder: AsyncEncoder<Disk::WriteDisk, T = T>,

pub async fn a_write_unload<U: Into<T>>( &mut self, new_value: U, ) -> Result<(), Coder::Error>

Available on crate feature async only.

Self::write_unload.

impl<T: Serialize + for<'de> Deserialize<'de> + Send + Sync, Disk: AsyncWriteDisk + AsyncReadDisk, Coder: AsyncEncoder<Disk::WriteDisk, T = T> + AsyncDecoder<Disk::ReadDisk, T = T>> BackedEntry<OnceCell<T>, Disk, Coder>

pub async fn a_mut_handle( &mut self, ) -> Result<BackedEntryAsyncMut<'_, Self>, <Coder as AsyncDecoder<Disk::ReadDisk>>::Error>

Available on crate feature async only.

Convenience wrapper for BackedEntryAsyncMut::mut_handle

impl<T: for<'de> Deserialize<'de> + Serialize + Send + Sync, Disk: AsyncReadDisk, Coder: AsyncDecoder<Disk::ReadDisk, T = T>> BackedEntry<OnceCell<T>, Disk, Coder>

pub async fn a_change_backing<OtherDisk, OtherCoder>( self, disk: OtherDisk, coder: OtherCoder, ) -> Result<BackedEntryAsync<T, OtherDisk, OtherCoder>, Either<Coder::Error, OtherCoder::Error>>
where OtherDisk: AsyncWriteDisk, OtherCoder: AsyncEncoder<OtherDisk::WriteDisk, T = T>,

Available on crate feature async only.

See Self::change_backing.

pub async fn to_sync<OtherDisk, OtherCoder, U, F, R>( self, disk: OtherDisk, coder: OtherCoder, blocking_fn: F, ) -> Result<BackedEntry<U, OtherDisk, OtherCoder>, Either<Coder::Error, OtherCoder::Error>>
where <Coder as AsyncDecoder<<Disk as AsyncReadDisk>::ReadDisk>>::Error: Send + Sync, U: Once<Inner = T>, OtherDisk: WriteDisk, OtherCoder: Encoder<OtherDisk::WriteDisk>, F: FnOnce(UpdateBlocking<U, OtherDisk, OtherCoder>) -> R, R: Future<Output = Result<BackedEntry<U, OtherDisk, OtherCoder>, OtherCoder::Error>>,

Available on crate feature async only.

Converts self into a synchronous backing.

Arguments

disk: Target synchronous disk

coder: Target synchronous disk

blocking_fn: A function to execute synchronous update inside.

blocking_fn

Can be used to move execution to a blocking thread. See crate::utils::blocking for convenience wrappers. Otherwise create a function that takes BlockingFn and uses BlockingFn::call to produce a result.

#[cfg(all(feature = "bincode", feature = "async_bincode", feature = "test"))] {
    use backed_data::{
        test_utils::cursor_vec,
        entry::{
            BackedEntryAsync,
            BackedEntryCell,
            formats::{
                AsyncBincodeCoder,
                BincodeCoder,
            },
        },
        utils::blocking::BlockingFn,
    };
    use tokio::runtime::Builder;
     
    const VALUES: &[u8] = &[1, 2, 5, 7];

    let rt = Builder::new_current_thread().build().unwrap();
    rt.block_on(async {
    cursor_vec!(backing);
    let mut disk: BackedEntryAsync<Box<[u8]>, _, AsyncBincodeCoder<_>> = BackedEntryAsync::with_disk(backing);
    disk.a_write_unload(VALUES).await;

    cursor_vec!(sync_backing);
    let sync_disk: BackedEntryCell<_, _, BincodeCoder<Box<[u8]>>> = disk.to_sync(
        sync_backing,
        BincodeCoder::default(),
        |f| async { f.call() }
    ).await.unwrap();
    });
}

impl<T: Serialize + for<'de> Deserialize<'de> + Sync + Send, Disk: AsyncWriteDisk, Coder: AsyncEncoder<Disk::WriteDisk, T = T>> BackedEntry<OnceCell<T>, Disk, Coder>

pub async fn from_sync<OtherDisk, OtherCoder, U, F, R>( other: BackedEntry<U, OtherDisk, OtherCoder>, disk: Disk, coder: Coder, blocking_fn: F, ) -> Result<Self, Either<OtherCoder::Error, Coder::Error>>
where U: Once<Inner = T>, OtherDisk: ReadDisk, OtherCoder: Decoder<OtherDisk::ReadDisk, T = T>, F: FnOnce(LoadBlocking<U, OtherDisk, OtherCoder>) -> R, R: Future<Output = Result<BackedEntry<U, OtherDisk, OtherCoder>, OtherCoder::Error>>,

Available on crate feature async only.

Converts from a synchronous backing into self.

Arguments

disk: Target synchronous disk

coder: Target synchronous disk

blocking_fn: A function to execute synchronous load inside.

blocking_fn

Can be used to move execution to a blocking thread. See crate::utils::blocking for convenience wrappers. Otherwise create a function that takes BlockingFn and uses BlockingFn::call to produce a result.

#[cfg(all(feature = "bincode", feature = "async_bincode", feature = "test"))] {
    use backed_data::{
        test_utils::cursor_vec,
        entry::{
            BackedEntryAsync,
            BackedEntryLock,
            formats::{
                AsyncBincodeCoder,
                BincodeCoder,
            },
        },
        utils::blocking::{
            tokio_blocking,
            BlockingFn,
        },
    };
    use tokio::runtime::Builder;
     
    const VALUES: &[u8] = &[1, 2, 5, 7];

    let rt = Builder::new_multi_thread().build().unwrap();
    rt.block_on(async {
    cursor_vec!(backing);
    let mut disk: BackedEntryAsync<Box<[u8]>, _, AsyncBincodeCoder<_>> = BackedEntryAsync::with_disk(backing);
    disk.a_write_unload(VALUES).await;

    cursor_vec!(sync_backing);
    let sync_disk: BackedEntryLock<_, _, BincodeCoder<Box<[u8]>>> = disk.to_sync(
        sync_backing,
        BincodeCoder::default(),
        |x| unsafe { tokio_blocking(x) }
    ).await.unwrap();
    });
}

impl<T: Once<Inner: Serialize>, Disk: WriteDisk, Coder> BackedEntry<T, Disk, Coder>
where Coder: Encoder<Disk::WriteDisk, T = T::Inner>,

pub fn update(&mut self) -> Result<(), Coder::Error>

Updates underlying storage with the current entry

pub fn write(&mut self, new_value: T::Inner) -> Result<(), Coder::Error>

Writes the new value to memory and disk.

See Self::write_unload to skip the memory write.

impl<T: Once<Inner: for<'de> Deserialize<'de>>, Disk: ReadDisk, Coder> BackedEntry<T, Disk, Coder>
where Coder: Decoder<Disk::ReadDisk, T = T::Inner>,

pub fn load(&self) -> Result<&T::Inner, Coder::Error>

Returns the entry, loading from disk if not in memory.

Will remain in memory until an explicit call to unload.

pub fn take(self) -> Result<T::Inner, Coder::Error>

Takes the inner value, loading from disk if not in memory.

impl<T: Once, Disk, Coder> BackedEntry<T, Disk, Coder>

pub fn is_loaded(&self) -> bool

impl<T: Once, Disk, Coder> BackedEntry<T, Disk, Coder>

pub fn unload(&mut self)

impl<T: Once<Inner: Serialize>, Disk: WriteDisk, Coder> BackedEntry<T, Disk, Coder>
where Coder: Encoder<Disk::WriteDisk, T = T::Inner>,

pub fn write_unload<U: Into<T::Inner>>( &mut self, new_value: U, ) -> Result<(), Coder::Error>

Write the value to disk only, unloading current memory.

See Self::write to keep the value in memory.

impl<T: Once<Inner: Serialize + for<'de> Deserialize<'de>>, Disk: WriteDisk + ReadDisk, Coder: Encoder<Disk::WriteDisk, T = T::Inner> + Decoder<Disk::ReadDisk, T = T::Inner>> BackedEntry<T, Disk, Coder>

pub fn mut_handle( &mut self, ) -> Result<BackedEntryMut<'_, Self>, <Coder as Decoder<Disk::ReadDisk>>::Error>

Convenience wrapper for BackedEntryMut::mut_handle

impl<T: Once<Inner: for<'de> Deserialize<'de> + Serialize>, Disk: ReadDisk, Coder: Decoder<Disk::ReadDisk, T = T::Inner>> BackedEntry<T, Disk, Coder>

pub fn change_backing<OtherDisk, OtherCoder>( self, disk: OtherDisk, coder: OtherCoder, ) -> Result<BackedEntry<T, OtherDisk, OtherCoder>, Either<Coder::Error, OtherCoder::Error>>
where OtherDisk: WriteDisk, OtherCoder: Encoder<OtherDisk::WriteDisk, T = T::Inner>,

Converts self to another disk and encoding representation.

This loads the value from the original disk in the original format (if necessary), and then encodes the value to the target disk. The original disk is not deleted. This may produce a disk read, and will always produce a disk write.

If switching between compatible coders (e.g. simd_json -> serde_json) without switching the backing disk, use Self::encoder_into to skip the extraneous disk reads/writes.

pub fn encoder_into<OtherCoder>(self) -> BackedEntry<T, Disk, OtherCoder>
where OtherCoder: From<Coder>,

Replaces self’s encoder without any disk operation.

pub fn encoder_try_into<OtherCoder>( self, ) -> Result<BackedEntry<T, Disk, OtherCoder>, OtherCoder::Error>
where OtherCoder: TryFrom<Coder>,

Replaces self’s encoder without any disk operation.

impl<T, Disk, Coder> BackedEntry<T, Disk, Coder>

pub fn get_disk(&self) -> &Disk

pub fn get_disk_mut(&mut self) -> &mut Disk

impl<T: Once, Disk, Coder> BackedEntry<T, Disk, Coder>

pub fn new(disk: Disk, coder: Coder) -> Self

See Self::with_disk when Coder implements default.

impl<T: Once, Disk, Coder: Default> BackedEntry<T, Disk, Coder>

pub fn with_disk(disk: Disk) -> Self

Self::new, but builds Coder from default.