Type Alias backed_data::entry::BackedEntryAsync

pub type BackedEntryAsync<T, Disk, Coder> = BackedEntry<OnceCell<T>, Disk, Coder>;

Available on crate feature async only.

Async specialization of BackedEntry.

Async versions of methods are prepended with a_.

Only implemented over a [tokio::sync::OnceCell], so it can be used within a tokio multi-threaded runtime without blocking.

Generics

T: the type to store.

Disk: an implementor of AsyncReadDisk and/or AsyncWriteDisk.

Coder: an implementor of AsyncEncoder and/or AsyncDecoder.

Aliased Type

struct BackedEntryAsync<T, Disk, Coder> { /* private fields */ }

Implementations

impl<T: Serialize + Send + Sync, Disk: AsyncWriteDisk, Coder> BackedEntryAsync<T, Disk, Coder>
where Coder: AsyncEncoder<Disk::WriteDisk, T = T>,

pub async fn a_update(&mut self) -> Result<(), Coder::Error>

See Self::update.

pub async fn a_write(&mut self, new_value: T) -> Result<(), Coder::Error>

See Self::write.

impl<T: for<'de> Deserialize<'de> + Send + Sync, Disk: AsyncReadDisk, Coder> BackedEntryAsync<T, Disk, Coder>
where Coder: AsyncDecoder<Disk::ReadDisk, T = T>,

pub async fn a_load(&self) -> Result<&T, Coder::Error>

See Self::load.

impl<T: Serialize + Send + Sync, Disk: AsyncWriteDisk, Coder> BackedEntryAsync<T, Disk, Coder>
where Coder: AsyncEncoder<Disk::WriteDisk, T = T>,

pub async fn a_write_unload<U: Into<T>>( &mut self, new_value: U, ) -> Result<(), Coder::Error>

Self::write_unload.

impl<T: Serialize + for<'de> Deserialize<'de> + Send + Sync, Disk: AsyncWriteDisk + AsyncReadDisk, Coder: AsyncEncoder<Disk::WriteDisk, T = T> + AsyncDecoder<Disk::ReadDisk, T = T>> BackedEntryAsync<T, Disk, Coder>

pub async fn a_mut_handle( &mut self, ) -> Result<BackedEntryAsyncMut<'_, Self>, <Coder as AsyncDecoder<Disk::ReadDisk>>::Error>

Convenience wrapper for BackedEntryAsyncMut::mut_handle

impl<T: for<'de> Deserialize<'de> + Serialize + Send + Sync, Disk: AsyncReadDisk, Coder: AsyncDecoder<Disk::ReadDisk, T = T>> BackedEntryAsync<T, Disk, Coder>

pub async fn a_change_backing<OtherDisk, OtherCoder>( self, disk: OtherDisk, coder: OtherCoder, ) -> Result<BackedEntryAsync<T, OtherDisk, OtherCoder>, Either<Coder::Error, OtherCoder::Error>>
where OtherDisk: AsyncWriteDisk, OtherCoder: AsyncEncoder<OtherDisk::WriteDisk, T = T>,

See Self::change_backing.

pub async fn to_sync<OtherDisk, OtherCoder, U, F, R>( self, disk: OtherDisk, coder: OtherCoder, blocking_fn: F, ) -> Result<BackedEntry<U, OtherDisk, OtherCoder>, Either<Coder::Error, OtherCoder::Error>>
where <Coder as AsyncDecoder<<Disk as AsyncReadDisk>::ReadDisk>>::Error: Send + Sync, U: Once<Inner = T>, OtherDisk: WriteDisk, OtherCoder: Encoder<OtherDisk::WriteDisk>, F: FnOnce(UpdateBlocking<U, OtherDisk, OtherCoder>) -> R, R: Future<Output = Result<BackedEntry<U, OtherDisk, OtherCoder>, OtherCoder::Error>>,

Converts self into a synchronous backing.

Arguments

disk: Target synchronous disk

coder: Target synchronous disk

blocking_fn: A function to execute synchronous update inside.

blocking_fn

Can be used to move execution to a blocking thread. See crate::utils::blocking for convenience wrappers. Otherwise create a function that takes BlockingFn and uses BlockingFn::call to produce a result.

#[cfg(all(feature = "bincode", feature = "async_bincode", feature = "test"))] {
    use backed_data::{
        test_utils::cursor_vec,
        entry::{
            BackedEntryAsync,
            BackedEntryCell,
            formats::{
                AsyncBincodeCoder,
                BincodeCoder,
            },
        },
        utils::blocking::BlockingFn,
    };
    use tokio::runtime::Builder;
     
    const VALUES: &[u8] = &[1, 2, 5, 7];

    let rt = Builder::new_current_thread().build().unwrap();
    rt.block_on(async {
    cursor_vec!(backing);
    let mut disk: BackedEntryAsync<Box<[u8]>, _, AsyncBincodeCoder<_>> = BackedEntryAsync::with_disk(backing);
    disk.a_write_unload(VALUES).await;

    cursor_vec!(sync_backing);
    let sync_disk: BackedEntryCell<_, _, BincodeCoder<Box<[u8]>>> = disk.to_sync(
        sync_backing,
        BincodeCoder::default(),
        |f| async { f.call() }
    ).await.unwrap();
    });
}

impl<T: Serialize + for<'de> Deserialize<'de> + Sync + Send, Disk: AsyncWriteDisk, Coder: AsyncEncoder<Disk::WriteDisk, T = T>> BackedEntryAsync<T, Disk, Coder>

pub async fn from_sync<OtherDisk, OtherCoder, U, F, R>( other: BackedEntry<U, OtherDisk, OtherCoder>, disk: Disk, coder: Coder, blocking_fn: F, ) -> Result<Self, Either<OtherCoder::Error, Coder::Error>>
where U: Once<Inner = T>, OtherDisk: ReadDisk, OtherCoder: Decoder<OtherDisk::ReadDisk, T = T>, F: FnOnce(LoadBlocking<U, OtherDisk, OtherCoder>) -> R, R: Future<Output = Result<BackedEntry<U, OtherDisk, OtherCoder>, OtherCoder::Error>>,

Converts from a synchronous backing into self.

Arguments

disk: Target synchronous disk

coder: Target synchronous disk

blocking_fn: A function to execute synchronous load inside.

blocking_fn

Can be used to move execution to a blocking thread. See crate::utils::blocking for convenience wrappers. Otherwise create a function that takes BlockingFn and uses BlockingFn::call to produce a result.

#[cfg(all(feature = "bincode", feature = "async_bincode", feature = "test"))] {
    use backed_data::{
        test_utils::cursor_vec,
        entry::{
            BackedEntryAsync,
            BackedEntryLock,
            formats::{
                AsyncBincodeCoder,
                BincodeCoder,
            },
        },
        utils::blocking::{
            tokio_blocking,
            BlockingFn,
        },
    };
    use tokio::runtime::Builder;
     
    const VALUES: &[u8] = &[1, 2, 5, 7];

    let rt = Builder::new_multi_thread().build().unwrap();
    rt.block_on(async {
    cursor_vec!(backing);
    let mut disk: BackedEntryAsync<Box<[u8]>, _, AsyncBincodeCoder<_>> = BackedEntryAsync::with_disk(backing);
    disk.a_write_unload(VALUES).await;

    cursor_vec!(sync_backing);
    let sync_disk: BackedEntryLock<_, _, BincodeCoder<Box<[u8]>>> = disk.to_sync(
        sync_backing,
        BincodeCoder::default(),
        |x| unsafe { tokio_blocking(x) }
    ).await.unwrap();
    });
}