backed_data
Module extra_docs

Additional description of the library.

No code is kept in this module.

Modules

alternatives

backed_data provides a way to store large data, but it is not the only way. Each section addresses an alternative approach with some broad pros and cons.

deps

Dependencies are broken down by feature category, then feature.

features

Features are broken into four category headers. The default features are bincode, directory, and zstd. With no features available, synchronous BackedEntry is available with plainfile disks, but the library needs at least one coder feature or custom coder implementation to provide the entry.

unsafe_usage

Unsafe code information.