Module backed_data::extra_docs::features

Features are broken into four category headers. The default features are bincode, directory, and zstd. With no features available, synchronous BackedEntry is available with plainfile disks, but the library needs at least one coder feature or custom coder implementation to provide the entry.

all_disks enables all disk types, and all_formats enables all coder formats.

Library Functionality

async

Provide asynchronous variants library-wide, usually methods prepended with a_. Enabled by any executor feature.

array

Provide BackedArray to split one array into multiple smaller BackedEntry arrays in storage, but represent as a single array to the user.

unsafe_array

Enable features for array that requires unsafe code. Needed for mutable iterators and access on containers that return handles instead of direct references (e.g. SecretVecWrapper).

directory

Provide DirectoryBackedArray to wrap BackedArray. This simplifies appending data to the backing arrays and saving/reloading arrays.

Async Executors

None of these features are necessary to use an async executor, but they provide a correct AsyncFile for the Plainfile disks to use. They also give convenience handles in blocking.

tokio

smol

Disk Types

Plainfile variant disks are available with no features. Some executor feature needs to be chosen in order for those base types to have asynchronous file access. Custom disks can be defined by implementing WriteDisk and ReadDisk (and/or the async versions). They can also be defined using crate::entry::disks::custom types.

all_disks

Enable all disk types.

encrypted - Encrypted

Uses secrets and aes-gcm to encrypt all data stored to disk, and minimize the risk of memory-peeking in the encryption/decryption process.

mmap - Mmap

Enables the use of mmap calls on UNIX systems (guarded by mmap_impl) to replace regular file I/O. Reuses mapped memory when possible to reduce reading from disk costs.

network - Network

Uses reqwest to read from a remote resource.

zstd - ZstdDisk

zstd-rs compression and decompression.

async_zstd - ZstdDisk

zstd, but asynchronous.

zstdmt/async_zstdmt - ZstdDisk

Enable zstd with multithreaded compression support. May degrade performance; test that the data compression speed actually benefits from this feature. This is not necessary to compress multiple files with zstd at once, only to use multiple threads for the compression of a single file.

Predefined Formats

Formats can be custom-defined by implementing Encoder and Decoder (and/or the async versions). These features give ready to use defaults.

(All formats use Serde, but the Serde traits do not implement over Read and Write).

all_formats

Enable all formats.

Synchronous Formats

bincode - BincodeCoder

Binary encoding format.

serde_json - SerdeJsonCoder

Serde’s flagship (de)serializer.

simd_json - SimdJsonCoder

Significantly faster than serde_json.

csv - CsvCoder

Allows for reading from/writing to csv files.

Async Formats

async_bincode - AsyncBincodeCoder

Binary format. NOT compatible with bincode, due to a size header.

async_csv - AsyncCsvCoder

Compatible with csv.