backed_data::utils::blocking
Function tokio_blocking

pub async unsafe fn tokio_blocking<'a, T, F>(arg: F) -> T
where
    T: Send + Sync,
    F: BlockingFn<Output = T> + Send,

Available on crate features async and tokio only.

Executes on a spawned tokio thread.

If the safety rules will be met, this can be converted to a safe future via a closure (unsafe {|x| tokio_blocking(x)}).

Safety

This uses a non-blocking async_scoped spawn. If arg has a non-static lifetime, this future CANNOT be forgotten (std::mem::forget) mid-execution, as that will invalidate arg for the spawned thread and possibly cause a panic.