Struct backed_data::utils::AsyncCompatCursor

pub struct AsyncCompatCursor<T>(pub Cursor<T>);

std::io::Cursor wrapper that provides async methods.

futures does not define a cursor for both async and sync read/write, so we do it ourselves. tokio provides the async read/write methods on std::io::Cursor directly, futures wraps std::io::Cursor and does not implement regular read/write on it.

This is almost a complete direct copy of futures::io::Cursor for the async methods, with inner swapped for 0.

Tuple Fields
0: Cursor<T>