Trait backed_data::utils::ProtectedAccess

pub trait ProtectedAccess {
    type Val;

    // Required methods
    unsafe fn get_mut_unsafe(&self) -> impl AsMut<Self::Val>;
    unsafe fn get_unsafe(&self) -> impl AsRef<Self::Val>;
    fn get_mut(&mut self) -> impl AsMut<Self::Val>;
    fn inner(self) -> Self::Val;
    fn wrap(val: Self::Val) -> Self;
}

Trait to combine Mutex (allowing panics) and UnsafeCell in one generic.

Mutex provides the thread-safe access pattern, UnsafeCell optimizes for single-threaded performance.

Required Associated Types

type Val

Required Methods

unsafe fn get_mut_unsafe(&self) -> impl AsMut<Self::Val>

Mutable reference return from an immutable context.

Intended for structs where mutation is used to set a single value and hold that value across multiple mutable calls.

Safety

This should only be used to update values in a way that does not invalidate any other live references. Borrow self mutably before any invalidations, to make sure no references exist.

unsafe fn get_unsafe(&self) -> impl AsRef<Self::Val>

Immutable reference return.

Safety

Requires Self::get_mut_unsafe is used properly to be valid.

fn get_mut(&mut self) -> impl AsMut<Self::Val>

Regular mutable reference return.

Safety

Requires Self::get_mut_unsafe is used properly to be valid.

fn inner(self) -> Self::Val

fn wrap(val: Self::Val) -> Self

Object Safety

This trait is not object safe.