backed_data
Module entry

Defines BackedEntry, the core of this library.

Also defines disks and formats, which define the backing for BackedEntry.

Re-exports

pub use adapters::SyncAsAsync;

async

pub use async_impl::*;

async

pub use sync_impl::*;

Modules

adaptersasync

Defines adapters that allow synchronous disks and/or format encoders to be used asynchronously.

async_implasync

disks

Defines the underlying storage for BackedEntry.

formats

Defines the encoding/decoding formats for backed disks.

sync_impl

Structs

BackedEntry

Entry kept on some backing storage, loaded into memory on usage.

Traits

BackedEntryTrait

Wrapper to minimize bounds specification for BackedEntry.

Type Aliases

BackedEntryArr

Thread-unsafe typedef of BackedEntry for arrays.

BackedEntryArrLock

Thread-safe typedef of BackedEntry for arrays.

BackedEntryAsyncasync

Async specialization of BackedEntry.

BackedEntryBox

Thread-unsafe typedef of BackedEntry for non-pointer types.

BackedEntryBoxLock

Thread-safe typedef of BackedEntry for non-pointer types.

BackedEntryCell

Thread unsafe BackedEntry with inner value T.

BackedEntryLock

Thread safe BackedEntry with inner value T.