Trait backed_data::entry::formats::AsyncEncoder

pub trait AsyncEncoder<Target: ?Sized> {
    type Error: From<Error>;
    type T: ?Sized + Serialize + Send + Sync;

    // Required method
    fn encode(
        &self,
        data: &Self::T,
        target: Target,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send;
}

Available on crate feature async only.

A format encoder that can be used asynchronously.

Implementation

It is the responsibility of trait implementors to call [AsyncWriteExt::flush] and [AsyncWriteExt::close] to ensure encodings write out to disk. Any implementor that does not to call these may fail to put all data on disk, making the encoding invalid and causing read failure.

Required Associated Types

type Error: From<Error>

type T: ?Sized + Serialize + Send + Sync

Required Methods

fn encode( &self, data: &Self::T, target: Target, ) -> impl Future<Output = Result<(), Self::Error>> + Send

Fully write out formatted data to a target disk.

Object Safety

This trait is not object safe.